requirements.txt
# To install the required packages, run:
# pip install -r requirements.txt
pygame
numpy
svg.path